refactor: new compaction code

+++ internal/dmap/janitor.go
@@ -15,6 +15,7 @@
 package dmap
 
 import (
+	"strings"
 	"time"
 
 	"github.com/buraksezer/olric/internal/cluster/partitions"
@@ -39,13 +40,20 @@ func wipeOutFragment(part *partitions.Partition, name string, f *fragment) error
 func (s *Service) deleteEmptyFragments() {
 	janitor := func(part *partitions.Partition) {
 		part.Map().Range(func(name, tmp interface{}) bool {
+			if !strings.HasPrefix(name.(string), "dmap.") {
+				// This fragment belongs to a different data structure.
+				return true
+			}
+
 			f := tmp.(*fragment)
 			f.Lock()
 			defer f.Unlock()
+
 			if f.storage.Stats().Length != 0 {
-				// Continue scanning.
+				// It's not empty. Continue scanning.
 				return true
 			}
+
 			err := wipeOutFragment(part, name.(string), f)
 			if err != nil {
 				s.log.V(3).Printf("[ERROR] Failed to delete empty DMap fragment (kind: %s): %s on PartID: %d",
@@ -53,6 +61,7 @@ func (s *Service) deleteEmptyFragments() {
 				// continue scanning
 				return true
 			}
+
 			s.log.V(4).Printf("[INFO] Empty DMap fragment (kind: %s) has been deleted: %s on PartID: %d",
 				part.Kind(), name, part.ID())
 			return true
@@ -68,7 +77,7 @@ func (s *Service) deleteEmptyFragments() {
 	}
 }
 
-func (s *Service) janitor() {
+func (s *Service) janitorWorker() {
 	defer s.wg.Done()
 	timer := time.NewTimer(s.config.DMaps.CheckEmptyFragmentsInterval)
 	defer timer.Stop()
